+++ web/app/(main-layout)/page.tsx
@@ -4,7 +4,6 @@ import CommandViewer from "@/components/interface/command-viewer";
 import EditorToolbar from "@/components/interface/editor-toolbar";
 import ExtensionSuggestionOverlay from "@/components/interface/extension-suggestion-overlay";
 import { EditorContext } from "@/components/providers/editor-context-provider";
-import Voice from "@/components/tools/voice";
 import ConsolePanelView from "@/components/views/console-panel/console-panel-view";
 import ViewArea from "@/components/views/view-area";
 import { useContext } from "react";
@@ -18,7 +17,6 @@ export default function HomePage() {
       <ViewArea />
       <ConsolePanelView />
 
-      <Voice />
       {false && <ExtensionSuggestionOverlay />}
       {editorContext?.editorStates.isCommandViewerOpen && <CommandViewer />}
     </div>