+++ web/components/interface/command-viewer.tsx
@@ -1,9 +1,18 @@
 import useCommands from "@/lib/hooks/use-commands";
-import { Command } from "@/lib/types";
-import { addToast, Button, Input, Listbox, ListboxItem } from "@heroui/react";
+import { Command, MenuAction } from "@/lib/types";
+import {
+  addToast,
+  Button,
+  Input,
+  Kbd,
+  Listbox,
+  ListboxItem,
+} from "@heroui/react";
 import { useCallback, useContext, useEffect, useRef, useState } from "react";
 import { EditorContext } from "../providers/editor-context-provider";
 import Icon from "../misc/icon";
+import usePlatformAIAssistant from "@/lib/hooks/use-platform-ai-assistant";
+import { useMenuActions } from "@/lib/hooks/use-menu-actions";
 
 const inputPlaceholders = [
   "Type anything...",
@@ -16,13 +25,15 @@ const inputPlaceholders = [
 export default function CommandViewer() {
   const editorContext = useContext(EditorContext);
 
+  const { chatWithAssistant } = usePlatformAIAssistant();
   const { commands, runCommand, setKeywordFilter } = useCommands();
+  const { registerMenuAction, unregisterMenuAction } = useMenuActions("view");
 
   const [inputPlaceholder, setInputPlaceholder] = useState("");
-
   const [selectCommandIndex, setSelectCommandIndex] = useState(-1);
-
-  const [inputValue, setInputValue] = useState("");
+  const [inputTextValue, setInputTextValue] = useState("");
+  const [isInputVoice, setIsInputVoice] = useState(false);
+  const [isOutputVoice, setIsOutputVoice] = useState(false);
 
   const runCommandCallback = useCallback(
     async (command: Command) => {
@@ -50,56 +61,114 @@ export default function CommandViewer() {
       inputPlaceholders[Math.floor(Math.random() * inputPlaceholders.length)],
     );
 
-    // Listen to keyboard events for up/down arrow keys
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === "ArrowUp") {
-        setSelectCommandIndex((prev) =>
-          prev === 0 ? commands.length - 1 : prev - 1,
-        );
-      } else if (event.key === "ArrowDown") {
-        setSelectCommandIndex((prev) =>
-          prev === commands.length - 1 ? 0 : prev + 1,
-        );
-      }
-    };
-    window.addEventListener("keydown", handleKeyDown);
-
     return () => {
       clearInterval(interval);
-      window.removeEventListener("keydown", handleKeyDown);
     };
   }, []);
-
   useEffect(() => {
-    // Run the selected command on enter key press
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === "Enter" && selectCommandIndex !== -1) {
-        runCommandCallback(commands[selectCommandIndex]);
-      } else if (event.key === "Escape" || event.key === "F1") {
-        event.preventDefault();
-        editorContext?.setEditorStates((prev) => ({
-          ...prev,
-          isCommandViewerOpen: false,
-        }));
-      }
-    };
-    window.addEventListener("keydown", handleKeyDown);
-    return () => {
-      window.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [selectCommandIndex, commands]);
-
-  useEffect(() => {
-    if (inputValue !== "") {
+    if (inputTextValue !== "") {
       // Assume commands are ordered by relevance.
       // Filter commands based on the input value
-      setKeywordFilter(inputValue);
+      setKeywordFilter(inputTextValue);
       // Choose the first command suggestion.
       setSelectCommandIndex(0);
     } else {
       setKeywordFilter(undefined);
     }
-  }, [inputValue]);
+  }, [inputTextValue]);
+
+  useEffect(() => {
+    handlePressedKeys(editorContext?.editorStates.pressedKeys ?? []);
+  }, [editorContext?.editorStates.pressedKeys]);
+
+  function handleKeyDown(e: KeyboardEvent) {
+    // Prevent default behavior for certain keys
+    const key = e.key;
+
+    if (editorContext?.editorStates.pressedKeys.includes(key)) {
+      return;
+    }
+
+    const keysToPrevent = [
+      "Enter",
+      "Escape",
+      "F1",
+      "Control",
+      "Meta",
+      "Alt",
+      "Shift",
+    ];
+    if (keysToPrevent.includes(key)) {
+      e.preventDefault();
+      // @ts-expect-error continuePropagation is not in the type definition
+      e.continuePropagation();
+    }
+
+    const keys = [...(editorContext?.editorStates.pressedKeys ?? []), key];
+    editorContext?.setEditorStates((prev) => {
+      // Prevent duplicate keys
+      if (!prev.pressedKeys.includes(key)) {
+        return {
+          ...prev,
+          pressedKeys: keys,
+        };
+      }
+      return prev;
+    });
+  }
+
+  function handleKeyUp(e: KeyboardEvent) {
+    const key = e.key;
+
+    const keysToPrevent = [
+      "Enter",
+      "Escape",
+      "F1",
+      "Control",
+      "Meta",
+      "Alt",
+      "Shift",
+    ];
+    if (keysToPrevent.includes(key)) {
+      e.preventDefault();
+      // @ts-expect-error continuePropagation is not in the type definition
+      e.continuePropagation();
+    }
+
+    const keys =
+      editorContext?.editorStates.pressedKeys.filter((k) => k !== key) ?? [];
+
+    editorContext?.setEditorStates((prev) => ({
+      ...prev,
+      pressedKeys: keys,
+    }));
+  }
+
+  function handlePressedKeys(pressedKeys: string[]) {
+    // Run the selected command on enter key press
+    const isEnterPressed = pressedKeys.includes("Enter");
+    const isArrowUpPressed = pressedKeys.includes("ArrowUp");
+    const isArrowDownPressed = pressedKeys.includes("ArrowDown");
+    const isControlPressed = pressedKeys.includes("Control");
+    if (isEnterPressed && isControlPressed && selectCommandIndex !== -1) {
+      // Run command if ctrl is pressed
+      console.log("Running command");
+      runCommandCallback(commands[selectCommandIndex]);
+    } else if (isEnterPressed && !isControlPressed) {
+      // Chat with assistant if ctrl is not pressed
+      console.log("Chatting with assistant");
+      chatWithAssistant(inputTextValue, isOutputVoice);
+    }
+    else if (isArrowUpPressed) {
+      setSelectCommandIndex((prev) =>
+        prev === 0 ? commands.length - 1 : prev - 1,
+      );
+    } else if (isArrowDownPressed) {
+      setSelectCommandIndex((prev) =>
+        prev === commands.length - 1 ? 0 : prev + 1,
+      );
+    }
+  }
 
   return (
     <div className="absolute top-20 left-1/2 z-50 -translate-x-1/2">
@@ -112,11 +181,29 @@ export default function CommandViewer() {
           }}
           label="Command Center"
           placeholder={inputPlaceholder}
-          value={inputValue}
-          onValueChange={setInputValue}
+          value={inputTextValue}
+          onValueChange={setInputTextValue}
           autoFocus
           endContent={
             <div className="flex h-full items-center justify-center">
+              {!inputTextValue && (
+                <div className="flex h-full items-center justify-center">
+                  <Button
+                    isIconOnly
+                    variant="light"
+                    onPress={() => {
+                      setIsInputVoice((prev) => !prev);
+                    }}
+                    size="sm"
+                  >
+                    {isInputVoice ? (
+                      <Icon name="mic" className="text-primary" />
+                    ) : (
+                      <Icon name="mic_off" variant="outlined" />
+                    )}
+                  </Button>
+                </div>
+              )}
               <Button
                 isIconOnly
                 variant="light"
@@ -126,13 +213,21 @@ export default function CommandViewer() {
                     isCommandViewerOpen: false,
                   }));
                 }}
+                size="sm"
               >
                 <Icon name="close" />
               </Button>
             </div>
           }
+          onKeyDown={(e) => handleKeyDown(e as any)}
+          onKeyUp={(e) => handleKeyUp(e as any)}
         />
         <div className="bg-content1 rounded-2xl shadow-md">
+          <div className="px-3 pt-2">
+            <p className="text-sm font-bold whitespace-nowrap">
+              Found {commands.length} commands
+            </p>
+          </div>
           <Listbox
             selectionMode="single"
             selectedKeys={
@@ -154,6 +249,13 @@ export default function CommandViewer() {
                 key={index.toString()}
                 className="data-[is-selected=true]:bg-primary/20"
                 data-is-selected={selectCommandIndex === index}
+                endContent={
+                  selectCommandIndex === index && (
+                    <div className="absolute right-7">
+                      <Kbd>Ctrl + Enter</Kbd>
+                    </div>
+                  )
+                }
               >
                 {command.commandInfo.name}
               </ListboxItem>