+++ web/lib/hooks/use-platform-ai-assistant.ts
@@ -4,19 +4,41 @@ import useTTS from "./use-tts";
 import { EditorContext } from "@/components/providers/editor-context-provider";
 import { getAPIKey } from "@/lib/settings/api-manager-utils";
 import { editorAssistantAgent } from "@/lib/agent/built-in-agents/editor-assistant";
-import { getAgentLLMConfig, runAgentMethod } from "@/lib/agent/agent-runner";
+import {
+  getAgentLLMConfig,
+  runAgentMethodCloud,
+  runAgentMethodLocal,
+} from "@/lib/agent/agent-runner";
 import { getPlatform } from "@/lib/platform-api/platform-checker";
-import { PlatformEnum } from "@/lib/types";
+import {
+  AppViewConfig,
+  CanvasViewConfig,
+  PlatformEnum,
+  TabView,
+} from "@/lib/types";
 import toast from "react-hot-toast";
 import { getDefaultLLMConfig } from "@/lib/modalities/utils";
 import { usePlatformApi } from "./use-platform-api";
 import useCommands from "./use-commands";
 import { useTabViewManager } from "./use-tab-view-manager";
+import { ViewModeEnum } from "@pulse-editor/shared-utils";
 
 export default function usePlatformAIAssistant() {
   const editorContext = useContext(EditorContext);
 
+  const isUseManagedCloud =
+    editorContext?.persistSettings?.isUseManagedCloud ?? true;
+
+  const [history, setHistory] = useState<
+    {
+      role: "user" | "assistant";
+      text: string;
+      audio: ReadableStream | undefined;
+    }[]
+  >([]);
+
   const { platformApi } = usePlatformApi();
+
   const { runSpeech2Speech, stopSpeech2Speech, isRunning } = useSpeech2Speech();
   const { readText, playAudio } = useTTS();
   const { runCommand, commands } = useCommands();
@@ -129,7 +151,7 @@ export default function usePlatformAIAssistant() {
         return;
       }
 
-      const { analysis }: { analysis: string } = await runAgentMethod(
+      const { analysis }: { analysis: string } = await runAgentMethodLocal(
         llmKey,
         editorAssistantAgent.LLMConfig,
         editorAssistantAgent,
@@ -151,7 +173,144 @@ export default function usePlatformAIAssistant() {
     processAssistantResult();
   }, [assistantResult]);
 
-  async function handleMicFinished() {
+  async function runCloudAssistant(
+    input: ReadableStream | string | undefined,
+    isOutputAudio: boolean,
+  ) {
+    if (!editorContext) {
+      return;
+    }
+
+    if (getPlatform() === PlatformEnum.VSCode) {
+      toast.error(
+        "Voice Chat is not supported in VSCode Extension. Please use other versions for Voice Chat.",
+      );
+      return;
+    }
+
+    if (!editorContext.editorStates.isRecording) {
+      const agent = editorAssistantAgent;
+      const methodName = "useExtensionCommands";
+
+      runSpeech2Speech(async (inputText: string) => {
+        const currentPath =
+          editorContext.persistSettings?.projectHomePath &&
+          editorContext.editorStates.project
+            ? editorContext.persistSettings?.projectHomePath +
+              "/" +
+              editorContext.editorStates.project
+            : undefined;
+
+        const projectDirTree = [];
+        if (currentPath) {
+          const gitIgnorePath = `${currentPath}/.gitignore`;
+          const gitIgnoreLines: string[] = [];
+          if (await platformApi?.hasPath(gitIgnorePath)) {
+            const gitIgnoreFile = await platformApi?.readFile(gitIgnorePath);
+            const gitIgnoreContent = await gitIgnoreFile?.text();
+            const lines = gitIgnoreContent?.split("\n") ?? [];
+            gitIgnoreLines.push(...lines.filter((line) => line.trim() !== ""));
+          }
+          const tree =
+            (await platformApi?.listPathContent(currentPath, {
+              include: "all",
+              gitignore: gitIgnoreLines,
+              isRecursive: true,
+            })) ?? [];
+          projectDirTree.push(...tree);
+        }
+
+        const config =
+          getAgentLLMConfig(agent, methodName) ??
+          getDefaultLLMConfig(editorContext);
+
+        if (!config) {
+          toast.error("No LLM config found for agent.");
+          return "No LLM config found for agent. Please configure the LLM in settings.";
+        }
+
+        // Pipe the LLM result to Speech2Speech
+
+        function getCommandsArgs() {
+          return commands.map((cmd) => ({
+            cmdName: cmd.commandInfo.name,
+            parameters: Object.entries(cmd.commandInfo.parameters).map(
+              ([key, value]) => ({
+                name: key,
+                type: value.type,
+                description: value.description,
+              }),
+            ),
+          }));
+        }
+
+        const tabView: TabView | undefined =
+          activeTabView?.type === ViewModeEnum.App
+            ? {
+                ...activeTabView,
+                // Remove dynamic commands to avoid sending too large payload.
+                // This is already included in the commands argument.
+                config: {
+                  ...(activeTabView?.config as any),
+                  dynamicCommands: undefined,
+                } as AppViewConfig,
+              }
+            : activeTabView?.type === ViewModeEnum.Canvas
+              ? {
+                  ...activeTabView,
+                  config: {
+                    ...(activeTabView?.config as CanvasViewConfig),
+                    appConfigs:
+                      (
+                        activeTabView?.config as CanvasViewConfig
+                      )?.appConfigs?.map((appConfig) => ({
+                        ...appConfig,
+                        // Remove dynamic commands to avoid sending too large payload.
+                        // This is already included in the commands argument.
+                        dynamicCommands: undefined,
+                      })) ?? [],
+                  } as CanvasViewConfig,
+                }
+              : undefined;
+
+        const result = await runAgentMethodCloud(config, agent, methodName, {
+          chatHistory: [],
+          userMessage: inputText,
+          activeTabView: tabView,
+          availableCommands: getCommandsArgs(),
+          projectDirTree: projectDirTree,
+        });
+
+        const {
+          suggestedCmd,
+          suggestedArgs,
+          suggestedViewId,
+          response,
+        }: {
+          suggestedCmd: string;
+          suggestedArgs: {
+            name: string;
+            value: any;
+          }[];
+          suggestedViewId: string;
+          response: string;
+        } = result;
+
+        console.log("Agent suggestion:", result);
+
+        setAssistantResult(result);
+
+        return response;
+      });
+    } else {
+      stopSpeech2Speech();
+    }
+  }
+
+  async function runLocalAssistant(
+    input: ReadableStream | string | undefined,
+    isOutputAudio: boolean,
+  ) {
     if (!editorContext) {
       return;
     }
@@ -223,25 +382,62 @@ export default function usePlatformAIAssistant() {
 
         setUserVoiceMessage(inputText);
         // Pipe the LLM result to Speech2Speech
-        // const stream = await llm.generateStream(inputText);
-        const result = await runAgentMethod(llmKey, config, agent, methodName, {
-          chatHistory: [],
-          userMessage: inputText,
-          activeTabView,
-          commands: commands.map((commandInTab) => ({
-            cmd: commandInTab.commandInfo.name,
-            parameters: Object.entries(commandInTab.commandInfo.parameters).map(
+
+        function getCommandsArgs() {
+          return commands.map((cmd) => ({
+            cmdName: cmd.commandInfo.name,
+            parameters: Object.entries(cmd.commandInfo.parameters).map(
               ([key, value]) => ({
                 name: key,
                 type: value.type,
                 description: value.description,
               }),
             ),
-            description: commandInTab.commandInfo.description,
-            viewId: commandInTab.viewId,
-          })),
-          projectDirTree: projectDirTree,
-        });
+          }));
+        }
+
+        const tabView: TabView | undefined =
+          activeTabView?.type === ViewModeEnum.App
+            ? {
+                ...activeTabView,
+                // Remove dynamic commands to avoid sending too large payload.
+                // This is already included in the commands argument.
+                config: {
+                  ...(activeTabView?.config as any),
+                  dynamicCommands: undefined,
+                } as AppViewConfig,
+              }
+            : activeTabView?.type === ViewModeEnum.Canvas
+              ? {
+                  ...activeTabView,
+                  config: {
+                    ...(activeTabView?.config as CanvasViewConfig),
+                    appConfigs:
+                      (
+                        activeTabView?.config as CanvasViewConfig
+                      )?.appConfigs?.map((appConfig) => ({
+                        ...appConfig,
+                        // Remove dynamic commands to avoid sending too large payload.
+                        // This is already included in the commands argument.
+                        dynamicCommands: undefined,
+                      })) ?? [],
+                  } as CanvasViewConfig,
+                }
+              : undefined;
+
+        const result = await runAgentMethodLocal(
+          llmKey,
+          config,
+          agent,
+          methodName,
+          {
+            chatHistory: [],
+            userMessage: inputText,
+            activeTabView: tabView,
+            availableCommands: getCommandsArgs(),
+            projectDirTree: projectDirTree,
+          },
+        );
 
         const {
           suggestedCmd,
@@ -269,8 +465,20 @@ export default function usePlatformAIAssistant() {
     }
   }
 
-  async function runAssistant(isUseVoice: boolean) {
-    handleMicFinished();
+  /**
+   * Run the assistant with the given input and output settings.
+   * @param input User input, can be audio (ReadableStream) or text (string).
+   * @param isOutputAudio Whether the output should be audio.
+   */
+  async function chatWithAssistant(
+    input: ReadableStream | string | undefined,
+    isOutputAudio: boolean,
+  ) {
+    if (isUseManagedCloud) {
+      runCloudAssistant(input, isOutputAudio);
+    } else {
+      runLocalAssistant(input, isOutputAudio);
+    }
   }
 
   function getAgentSuggestedCommands(query: string, k: number) {
@@ -278,6 +486,8 @@ export default function usePlatformAIAssistant() {
   }
 
   return {
-    runAssistant,
+    isUseManagedCloud,
+    history,
+    chatWithAssistant,
   };
 }